Basic Combat v2
Added the Card object, which for now only holds the name of a card. Added lists of Card objects called deck, hand, and discard pile. Added a function to shuffle the deck and draw cards from the top, and an unused function to add cards from the discard pile into the deck. Running the game now shuffles the 5 blue and 5 red cards into the deck, draws 5 cards, and displays them on screen, and then prints out the contents of the deck, hand, and discard pile.