Basic Combat v4
Added the Card Display Prefab and the Hand GameObject, which are used by the CombatController to dynamically display the cards in the player's hand. Now you can draw a variable number of cards using DrawCards() at the start of the CombatController script. However, the player's hand position is not centered on the screen, and some code is left uncommented, but I have to go for now.