Basic Combat v7
Added touch support! When the player touches the screen, their all of the objects their touch overlaps with is returned, and one is chosen to be interacted with. It is easy to add a new type of interactable object to this process. Added the DisplayCard script as a means to access the CardData object from a clicked DisplayCard. Added functions to CombatController to allow the player to interact with the Deck and Discard Pile.