Added click-and-drag functionality to cards on screen
When the player clicks a card, they can now hold the mouse down to drag the card around the screen. When the player releases the mouse, the card is "played" and goes into the discard pile like before. New sample text has been added for when the player plays a card. This text is temporary and will eventually be replaced by the card effect execution functions. Added some functions and variables to TouchHandler to operate the click-and-drag functionality. Removed code processing player touch on a mobile device, because I don't expect we will need to process more than one touch at a time, and the Input.mousePosition, Input.GetMouseDown, etc functions can work equally well on mobile and pc.