Added the Catalog
Added the Catalog, which stores the data on all cards in the game. These cards can be added and edited from the inspector, and each card is represented by the new PlatonicCard script. The name is temporary, but the functionality is not! A PlatonicCard represents one unique card in the catalog. A specific card in the game still uses the CardData script, but now the CardData script does not contain information on its own and instead references the PlatonicCard it is based off of. Updated all relevant references to CardData to now point towards the associated PlatonicCard.