Added health and shields
Added an int that tracks the player's health, and another for shields. Playing a blue card adds 3 shields, and then displays the total number of shields on screen. At the end of the turn, the accumulated shields are all removed. The player's current health and max health are both displayed on a health bar. There is currently no way to change the player's health yet, but the updating function works correctly. Also added 3 sprites denoting the 3 ally characters. They do nothing, except contain some text and look pretty.