attack and defeat enemies
if a red card is placed on an enemy, the DisplayCard runs the DealDamageToEnemy function in the CombatController file. This can deal damage to that enemyData, and, if the damage is enough, defeat the enemy, which removes them from the game.

DisplayEnemies function now removes all current enemies before running its normal functionality.