Improved playing cards against enemies
Cleaned up some code. Also, added the "Requires Target" property to the PlatonicCard script. Now, when a card is released, it checks if the card requires a target and responds accordingly. Added the PlayCardWithTarget and PlayCardWithoutTarget functions to DisplayCard. These will likely get moved to another script as development continues.