Added dynamic effects!
Removed PlayCardWithTarget and PlayCardWithoutTarget, in favor of PlayCard, which takes an optional argument of a target. If the card requires a target or if it does not, either way, PlayCard gets called.

Added a list of strings to the PlatonicCard script to denote all of the individual effects of the card to be executed in sequence. So far, "do damage" and "shield" are supported, but mutliple effects in a row are allowed. To test, the Purple Card has been added, with the effect of doing 2 damage and then adding 2 shields. Effects added in the inspector that require a numerical value to execute are denoted as "do damage=4" or "shield=3", such that the name of the effect and the required value are seprated by an equals sign. This will likely change.