Improved the visuals for Cards
Added a new frame for cards! This is grey and comes from the July Humble Bundle. The DisplayCard prefab has been updated with this new frame style, and consequentially, the card-specific frames have been removed from PlatonicCard. Modified the TouchHandler script to elevate the Sorting Order of a DisplayCard while you are holding it, and return the DisplayCard to its usual Sorting Order after you let go. Overall, now the cards look better, and when you interact with them they move to the forefront of the screen.