Added support for multiple enemies
Now you can have up to 4 enemies in one combat. Their positions are set by looking at objects already in the scene when the game starts. There are 4 enemy objects in the scene, Enemy 1 through Enemy 4, and their positions are recorded when the combat starts. Those positions are referenced when an enemy needs to be added to the scene.