EnemySprite objects are updated instead of being replaced
Added the UpdateEnemyHP and UpdateEnemyAttacks functions, which display the health for a single enemy and the attacks for all enemies. This lets us keep an enemy on the screen for the whole combat, and we update the hp and attack strings instead of rendering a new enemy. Added an enemySprite reference to the EnemyData script.