Merged EnemyData and EnemySprite
Because we will always have exactly one EnemyData for each EnemySprite, those two scripts have been merged into Enemy. CombatController instantiates the Enemy prefab several times, and for each, positions it and assigns the associated Enemy script the appropriate data. Instance-specific variables, like total amount of damage taken, is now an attribute of the Enemy script. Additionally, now the enemies and cards used in a combat are designated from the CombatController script. When the scene starts, the names of cards and ids of enemies are read from a list, which is editable in the inspector. Also updated some comments.