Added Win and Loss functionality
Added the Win Canvas and Lose Canvas. Each has an associated script, which will eventually hold more functions to update visuals on a win or loss. At the start of the scene, the win and loss scripts hide all of their children. When the player wins or loses the combat, the children are shown again.