Added the Main Canvas script and gameobject
Added the Main Canvas script and gameobject, which will house most interactables for the combat scene. Removed the HealthDisplay gameobject from the combat scene world space, and moved it to the Main Canvas instead. The HealthDisplay uses a RectTransform to set its position relative to the screen corners. The Main Canvas gameobject has its UI Scale Mode set to "Scale With Screen Size" so that all UI elements keep their relative positions while scaling dependent on the resolution of the user's device.