Moved enemies to the Main Canvas
The visual display of the Enemies has been moved to the Main Canvas. The old Enemy prefab has been removed, and has been replaced with a new Enemy prefab that is supposed to be instantiated as a child of a canvas, instead of withing world space. Changed some code in CombatController and TouchHandler to accomodate this change. Currently, these enemies all have the same size sprite, no matter what the enemy is. The code that repositions the name, hp, and next attack text has been temporarily disabled.