Changed enemy rendering from dynamic to static
Added the Enemy Groups gameobject to the combat scene, which contains Small Enemies, Large Enemies, and Mixed Enemies gameobject. Each of these is a container for 2-4 objects representing enemies. These enemies have a fixed size, and are either small or large. Depending on which enemy sizes you add to the combat, the enemies are displayed using one of the 3 defined groups. The enemy display and data is no longer generated at runtime, so we do not need an Enemy prefab any longer. Modified the Start and AddEnemy functions in CombatController to make these changes possible.