Dynamically size and position cards in the hand
Cards are no longer removed every time their rendering position or display needs to change. Instead, we create a new DisplayCard object every time a card gets added to the hand. Whenever we need to draw, play, or discard a card, the DisplayCard's reference is remembered, and the card is moved to its new position. When the card needs to be discarded or the turn ends, the DisplayCards are deleted and their references are removed. The function to set the DisplayCard positions in the hand also changed. Now, there is a Hand Size gameobject with a predetermined width set up in the Hierarchy, and the DisplayCards are spaced equally to fill that width. Additionally, there is an "Ideal Hand Size" int provided to the CombatController in the inspector. If there are fewer cards in hand than that provided number, the cards do not spread out to fill the whole hand width. Instead, they are evenly spaced as if they had the provided number of cards in hand. So if the Ideal Hand Size is 5 cards and there are only 4 cards in hand, the spacing between cards is caclulated as if there were still 5 cards, and then that spacing is applied to the 4 cards in hand. To accomplish all this, the CreateDisplayCard and PositionCardsInHand functions were added, and the Start, DrawCards, MoveCardFromHandToDiscard, DiscardHand, and EndTurn functions were updated slightly.