Added the Overworld scene and the StaticVariables script
Added the Overworld scene. It currently is mostly empty, except for a few buttons that take the player to the Combat Scene. Each button provides a different list of enemies to fight. These enemies are taken as a string of the form 0-0-0-0, specifically each number is an enemy id, separated by a hyphen. This string is parsed into a list of integers, which is saved to the new StaticVariables script. Once the Combat scene loads, the saved enemy ids are read from StaticVariables and applied to the combat encounter.