When the player wins an encounter, they can choose a card to add to their deck
Added the EncounterDetails script, which holds data about which enemies the player will fight in a specified combat encounter, as well as the cards they can choose between when they win. This EncounterDetails variable is stored in StaticVariables, and referenced in CombatController. The card catalog, enemy catalog, and sprites for the numbers used in game have been moved from CombatController to OverworldThingy, and are stored in StaticVariables during runtime. Added 4 new temp cards for playtesting. The display data for each card reward is shown on the win screen, by editing 4 predefined UI gameobjects. Clicking one of these objects runs some new code in TouchHandler, which adds the chosen card to the player's deck and returns them to the overworld.