Improved card-playing functionality
Added the Card Dead Zone gameobject as a child of the Hand gameobject. Card Dead Zone is the only gameobject in the hand that is not deactivated when the combat scene loads. Card Dead Zone currently takes up the same amount of space in the hand as the Hand Size gameobject, but stretches across the whole screen instead. When the player releases a non-targeting card, and the center of the card is still in the Card Dead Zone, do not play the card, and return it to its prior position in the hand. Also, when a card is returned to its prior position, it is also returned to its prior place in the hierarchy. This place is caulcuated in CombatController.PositionCardsInHand, and is a new prioerty of the DisplayCard class.