Card effects can be selected from a drop-down in the inspector
Added the EffectBit class, which represents one segment of a card's effect, such as "deal 2 damage" or "gain 3 shield". The card effect type is enumerated in the Catalog, and an individual EffectBit's effect type can be chosen from a drop-down menu in the inspector. Overall, this change means that card effects are no longer provided and interpreted as strings, and instead as elements of the Catalog.EffectType enumerator.