Card rewards for an encounter are now random
Added Catalog.GetRandomCards, which returns a specified number of cards, chosen at random, from the list of all cards in the catalog. All cards returned are different. Card rewards for an encounter are no longer provided in the inspector. Added a random number generator variable to StaticVariables, called StaticVariables.random.