Added Small and Large Enemy prefabs
Added the Small Enemy and Large Enemy prefabs. Updated all small and large enemies in the combat scene to be instances of these prefabs. We will be making changes in the near future that will require all enemy objects to be updated consistently with one another, hence the need for the prefabs.