Added enemy Vulnerability.
Changed the "orange" card to deal Vulnerability to an enemy. A new enum called StatusEffects has been added to EnemyCatalog, which currently only contains the Vulnerable status. Playing "orange" on an enemy will give them the Vulnerable status. A Buff bar has been added underneath each enemy's attack text, and when the enemy recieves a buff/debuff, the buff bar changes to show the status. Currently, the Vulnerable status does not go away as turns pass, but that will change soon. When the enemy takes damage from a card and that enemy is Vulnerable, they take an extra 50% damage.