Enemy status conditions will now count down
Added the EnemyStatus class, which holds the status type and number of turns remaining for a single enemy status condition. Applying the Vulnerable status to an enemy now applies it for the specified number of turns. If an enemy is already Vulnerable, applying the same status will increase its duration. After enemies attack, all enemy status durations are now decremented, and removed entirely if the turn counter reaches zero. For now, enemy statuses only work for small enemies, to be fixed shortly.