Added PlatonicStatus
Added the PlatonicStatus class, which contains data about one enemy status condition type. Instances for the PlatonicStatus class are added to the EnemyCatalog script in the inspector, and contain the relevant EnemyStatus enumerator, the status' icon, and its text description. The status display for an enemy has been updated to use the PlatonicStatus icon instead of a pre-defined color.