Changed tap registration
Changed some of the minor details of how tap registration works. When the player releases their tap, it only shows the popup if their finger is over a card. Additionally, a touch release registers as a tap if it is within a specified radius of the tap-starting-position, so there is some room for user error now. The touch radius is currently 5% of screen width, although that can change easily later.