Added Allies
Added the AllyCatalog, Ally, and PlatonicAlly classes. These behave similarly to the implementation of enemies and cards. Currently, Allies contain a name, a sprite, and a list of cards that the player starts with when they have the specified ally in their party. Added three new Allies: Civilian, Rogue, and Medic, and several cards that those allies start with. No new card mechanics have been added. The images and names of your allies are displayed during a combat encounter.