Added the Tavern Scene and the Brawler Ally
Added the Tavern scene. This is now the first scene to load, and lets you choose your allies before starting a game. 3 columns of buttons are displayed in the Tavern, and each column contains one button for each possible ally. Clicking a button will set that ally to be in the designated spot in your group of allies. The colors of the buttons also updates to show which allies are currently selected. Added the Brawler ally, and some new cards to go with them. Renamed the OverworldThingy script to Overworld. Commented and cleaned up some code. Also the enemy's attack text updates to accomodate any Weak status they may or may not have.