Improved some screen scaling, and card details behavior
Adjusted the scroll size of the Ally selection. In the Combat scene, changed the size and position of the card details popup. Additionally, now the card details popup scales with screen size. Improved the card name text display for multiple screen sizes.

When the player taps a card during a combat encounter, show that card's details. Now, the next tap on the screen, no matter where it is, will not interact with anything, and instead will hide the card details popup.