Improved the card details popup again
A new Highlight image has been added to DisplayCard. When a card is tapped to show on the details popup, the highlight image is shown around the original card in the player's hand. Currently, the highlight is just a purple border around the card. When the player taps again to hide the card popup, the highlight disappears as well. Also, when the player taps to hide the card popup, the card returns to its original sorting layer spot in the hand, similar to when the player tries and fails to play a card for whatever reason.