Improved Ally Selection
Added the Ally Selection prefab, which contains the image, and name of an ally, as well as a few buttons. Several Ally Selection prefabs are instantiated in the Tavern script, inside the ally scroll views. Each instantiation corresponds to a single ally option. The Ally Selection buttons are provided functions during instatiation. One button selects the ally for play like previously, and the other button displays the ally's starting cards in a pop-up on screen. There is a third button that currently goes unused, but will eventually display the ally's sequence of actions.