Added the Deck and Discard Pile contents popup
Clicking on the deck or the discard pile will now show a new popup canvas, called PileDetailsPopup. This canvas contains a scroll view, which is populated with all of the cards in the specified pile arranged in rows. The cards cannot be interacted with. The contents of the chosen pile are displayed in a shuffled order. There is a button in the corner of the popup that will close it. While this popup is open, the player cannot interact with any other object on screen.