Added the bossFight bool to PlatonicEncounter
Each encounter is now specified as a boss fight or a non boss fight. Entering a room with the BOSS type will pick a random encounter with the boss tag as true. Entering a room with the COMBAT type will pick a random encounter with the boss tag as false. Added more encounters of both types. Added some comments. Changed the background color of the Rest room.