Added hallways in between rooms
Added the Hallway prefab. Each parent-child relationship between rooms now creates a hallway connecting those two rooms. The hallways are positioned and angled to connect between the centers of the rooms. The hallways are colored to show if the player can take that path, similarly to how the room buttons are colored.