Added the Fade Canvas, and improved the Overworld and Rest scenes
Added the Fade Canvas gameobject and FadeCanvas script. A Fade Canvas gameobject has been added to every scene, and controls the transparency of a black overlay that covers the whole screen. When transitioning to and from scenes, a fade-in or fade-out function is called from within the FadeCanvas script, and then the fading and scene transition is handled by the FadeCanvas. Added a player health display to the Overworld and Rest scenes. Added a "Gain Max Health" button and functionality to the Rest scene.