Added enemy damage and death animations
Added a slash animation that appears when an enemy takes damage.

Added an enemy death animation, which currently is just increasing the transparency of the enemy image and all attached text. The enemy death animation plays at the same time as the slashing animation for the attack that kills it. Once the last enemy dies, the win popup window is now delayed until after the enemy death animation is finished.