Added Tweening
Added tweening with DOTween. The first tweening implementation deals with cards in the hand. Playing a card now shrinks it, increases the transparency of a circular overlay on top of the card, and then moves the card to the discard pile before deleting it. When a card cannot be played and must return to its prior spot in the hand, the tweening is used to quickly move the card back into position.