Added the card queue
When a card is played, it goes onto the card queue. The combatcontroller.update checks if the current card at the top of the queue is not currently being played, and if it is, play it. The card will then do its entire effect in sequence, move to the discard, and destroy itself. Then, combatcontroller.update will play the next card in the queue.