Added permanent and temporary card removal
You can now choose to trash a card when you take a rest, which prompts a window of all of the cards in your deck that you can choose from to trash. Added the "remove from deck" card effect, which will remove the card with that effect from your deck until the end of the combat encounter.