Added the Shop
Added the Shop scene, which includes 8 unique cards randomly chosen from the catalog that the player can buy. Currently, cards have no actual cost to purchase. Purchased cards are tweened to the deck in the Shop and added to the player's deck. The shop includes a deck that the player can tap to see all cards in their deck. The shop can be added to a dungeon floor just like the other rooms.