Big Tavern Update!
The Tavern has been entirely changed, now showing the possible allies spread throughout the tavern. The currently-selected allies are visible in the top-left corner of the screen. By default, no allies are selected, and the player can tap an ally in the tavern to add them to the party. The party is now limited to one of each type of ally. The party member icons in the top left can be tapped to show the ally's card details, and also provide an option to remove the ally from the party, opening up room for a different ally to be chosen.

Fixed a bug preventing "damage all enemies" from working, and another bug which stopped the player from proceeding to the next floor after beating a boss room.