Added the first card Tag
Cards can now have a Tag, which is chosen from a pre-selected list in the Card Catalog. The first tag, "Beast", has been implemented. A card's tag is visible at the bottom of the card, if they have a tag. Cards without tags look the same as before. At the start of the tavern scene, the Card Catalog is iterated through, and a list is made of all cards with each tag, to iterate through later.

Added the Evoke, and Shape Change, along with several Beast cards. Evoke and Shape Change use a new mechanic, creating a new card with a specified tag and specified cost and adding it to the discard pile.