Added basic events
Events can now be added in the EventCatalog from the Tavern scene. Events have a text description of the scenario, several choices that appear depending on which allies the player has in the party, a description of the consequence of each choice, and an optional card reward for each choice. As of right now, each choice only has one reward, and the rewards can only be cards.