Added the "End Turn" card effect
Playing a card with the End Turn effect will prevent further cards from being added to the queue, send all cards in the current queue back to the player's hand, and then immediately end the player's turn as normal. Added the Haymaker card that uses this effect. Burn no longer decays at the end of the turn.